combined linkedlist and node into queue class
