--- lib/pool/node.js
@@ -1,11 +1,0 @@
-/** 
- * Singly linked list node 
- */
-export default class Node {
-    data;
-    next;
-
-    constructor(data) {
-        this.data = data;
-    }
-}
