+++ lib/pool/queue.js
@@ -1,11 +1,13 @@
-import LinkedList from './linkedlist.js';
+import EmptyQueueError from '../error/emptyqueueerror.js';
 
 /**
- * Partial Implementation of a standard queue based on the Queue ADT
+ * Partial implementation of a standard queue based on the Queue ADT
  *  - Backed by singly linked list
  */
 export default class Queue {
-    #linkedList;
+    #head;
+    #tail;
+    #size;
 
     /**
      * Gets the size of the queue.
@@ -13,7 +15,7 @@ export default class Queue {
      * @returns {Number}
      */
     get size() {
-        return this.#linkedList.size;
+        return this.#size;
     }
 
     /**
@@ -25,29 +27,50 @@ export default class Queue {
         return !this.size;
     }
 
-    constructor(array) {
-        this.#linkedList = LinkedList.fromArray(array);
+    constructor() {
+        this.#head = null;
+        this.#tail = null;
+        this.#size = 0;
+    }
+
+    /**
+     * Creation method to build a queue from an array.
+     * Maintains order of original array.
+     * 
+     * @param {Array} array
+     */
+    static fromArray(array) {
+        const instance = new Queue();
+        array.forEach(item => instance.enqueue(item));
+        return instance;
     }
 
     /**
      * Inserts an element at the rear of the queue.
      * 
-     * @param {any} data - element to enqueue
+     * @param {any} data
      */
     enqueue(data) {
-        this.#linkedList.insertLast(data);
+        const node = { data: data, next: null };
+        if (this.#tail) this.#tail.next = node;
+        this.#tail = node;
+        this.#head = this.#head ?? this.#tail;
+        this.#size++;
     }
 
     /**
-     * Removes an element from the front of the queue and returns it.
+     * Removes and returns the element from the front of the queue.
      * If queue is empty, an error is thrown.
      * 
      * @returns {any}
      */
     dequeue() {
-        if (this.empty) throw new EmptyCollectionError();
-        const data = this.#linkedList.head.data;
-        this.#linkedList.removeFirst();
+        if (this.empty) throw new EmptyQueueError();
+        const { data, next } = this.#head;
+        this.#head.next = null;
+        if (this.#head === this.#tail) this.#tail = null;
+        this.#head = next;
+        this.#size--;
         return data;
     }
 }
