+++ test/pool/queue.spec.js
@@ -1,6 +1,6 @@
 import assert from 'node:assert/strict';
 import Queue from '../../lib/pool/queue.js';
-import EmptyCollectionError from '../../lib/error/emptycollectionerror.js';
+import EmptyQueueError from '../../lib/error/emptyqueueerror.js';
 import { array } from '../helpers.js';
 
 /**
@@ -12,55 +12,81 @@ import { array } from '../helpers.js';
 const populateQueue = (queue, arr) => arr.forEach(i => queue.enqueue(i));
 
 describe('Queue', function() { 
-    let queue;
+    context('creation', function() {
+        describe('#fromArray', function() {
+            it('should maintain the order of the input array', function() {
+                const queueFromArray = Queue.fromArray(array.populated);
+                const queueAsArray = (() => {
+                    const elements = [];
+                    while(!queueFromArray.empty) elements.push(queueFromArray.dequeue());
+                    return elements;
+                })();
+                assert.deepEqual(queueAsArray, array.populated);
+            });
 
-    beforeEach(function() {
-        queue = new Queue();
-    });
+            it('size should increase by the length of the input array', function() {
+                assert.equal(Queue.fromArray(array.populated).size, array.populated.length);
+            });
 
-    describe('#enqueue', function() {
-        it('should insert value at the end', function() {
-            const expectedLastElement = 0;
-            queue.enqueue(expectedLastElement);
-            const lastElement = (() => {
-                let curr;
-                while(!queue.empty) curr = queue.dequeue();
-                return curr;
-            })();
-            assert.equal(lastElement, expectedLastElement);
-        });
-
-        it('size should increase by 1', function() {
-            const previousSize = queue.size;
-            queue.enqueue(0);
-            assert.equal(queue.size, previousSize + 1);
+            it('should return an empty queue if input array is empty', function() {
+                assert.equal(Queue.fromArray(array.empty).size, 0);
+            });
         });
     });
 
-    describe('#dequeue', function() {
-        context('queue is empty', function() {
-            it('should throw error', function() {
-                const expectedError = new EmptyCollectionError();
-                // This bound to instance method because instance reference 
-                // is not passed in with function reference
-                assert.throws(queue.dequeue.bind(queue), expectedError);
-            });
+    context('operation', function() {
+        let queue;
+
+        beforeEach(function() {
+            queue = new Queue();
         });
-        
-        context('queue is not empty', function() {
-            beforeEach(function() {
-                populateQueue(queue, array.populated);
+
+        // What about when list is empty?
+        describe('#enqueue', function() {
+            it('should insert value at the end', function() {
+                const expectedLastElement = 0;
+                queue.enqueue(expectedLastElement);
+                const lastElement = (() => {
+                    let curr;
+                    while(!queue.empty) curr = queue.dequeue();
+                    return curr;
+                })();
+                assert.equal(lastElement, expectedLastElement);
             });
 
-            it('should return the value at the front', function() {
-                const data = queue.dequeue();
-                assert.equal(data, array.populated[0]);
-            });
-
-            it('size should decrease by 1', function() {
+            it('size should increase by 1', function() {
                 const previousSize = queue.size;
-                queue.dequeue();
-                assert.equal(queue.size, previousSize - 1);
+                queue.enqueue(0);
+                assert.equal(queue.size, previousSize + 1);
+            });
+        });
+
+        describe('#dequeue', function() {
+            context('queue is empty', function() {
+                it('should throw error', function() {
+                    const expectedError = new EmptyQueueError();
+                    // This bound to instance method because instance reference 
+                    // is not passed in with function reference
+                    assert.throws(queue.dequeue.bind(queue), expectedError);
+                });
+            });
+            
+            // What about when size is 1?
+            context('queue is not empty', function() {
+                beforeEach(function() {
+                    populateQueue(queue, array.populated);
+                });
+
+                it('should return the value at the front', function() {
+                    const data = queue.dequeue();
+                    assert.equal(data, array.populated[0]);
+                });
+
+                it('size should decrease by 1', function() {
+                    const previousSize = queue.size;
+                    queue.dequeue();
+                    assert.equal(queue.size, previousSize - 1);
+                });
             });
         });
     });
